projectFiles - basic tests and scaffold specifics for local / s3. same for sequence
