+++ test/server/data/persistence/persistence.spec.js
@@ -17,7 +17,6 @@ import {
 } from '../../../../server/utils/fileSystem';
 import Project from '../../../../src/models/Project';
 import Block from '../../../../src/models/Block';
-import { validPseudoMd5, generatePseudoMd5, parsePseudoMd5 } from '../../../../src/utils/sequenceMd5';
 
 import * as filePaths from '../../../../server/utils/filePaths';
 import * as versioning from '../../../../server/data/versioning';
@@ -489,58 +488,6 @@ describe('Server', () => {
             .then(() => versioning.log(projectRepoDataPath))
             .then(log => {
               expect(log.length).to.equal(versionLog.length);
-            });
-        });
-      });
-
-      describe('sequence', () => {
-        const rangeStart = 20;
-        const rangeEnd = 35;
-        const sequence = 'ACTAGCTAGCTAGCTGACTAGCTAGCTGATCGTAGCGATCTACTGATCAGCTACTGTACGTACGTGACTG';
-        const rangedSequence = sequence.substring(rangeStart, rangeEnd);
-
-        const realMd5 = md5(sequence);
-        const pseudoMd5 = generatePseudoMd5(realMd5, rangeStart, rangeEnd);
-
-        const filePath = filePaths.createSequencePath(realMd5);
-
-        //skip test suite if not using s3
-        before(function() {
-          if (s3.useRemote) {
-            this.skip();
-          }
-        });
-
-        it('sequenceWrite() should write a sequence', () => {
-          return persistence.sequenceWrite(realMd5, sequence)
-            .then(() => fileRead(filePath, false))
-            .then(read => {
-              expect(read).to.equal(sequence);
-            });
-        });
-
-        it('sequenceRead() should read a sequence', () => {
-          return fileRead(filePath, false)
-            .then(fileResult => {
-              assert(fileResult === sequence, 'sequence should be written already');
-
-              return persistence.sequenceGet(realMd5)
-                .then(getResult => {
-                  expect(getResult).to.equal(fileResult);
-                  expect(getResult).to.equal(sequence);
-                });
-            });
-        });
-
-        it('sequenceRead() should read a sequence when md5 is specifying a range', () => {
-          return fileRead(filePath, false)
-            .then(fileResult => {
-              assert(fileResult === sequence, 'sequence should be written already');
-
-              return persistence.sequenceGet(pseudoMd5)
-                .then(getResult => {
-                  expect(getResult).to.equal(rangedSequence);
-                });
             });
         });
       });
